move components to seperate files
