+++ src/pages/api/schedule/[school].ts
@@ -86,8 +86,6 @@ export default function handler(req: NextApiRequest, res: NextApiResponse) {
 
 		return res.status(200).json(scheduleDB);
 	} catch (error) {
-		console.error(error);
-		console.log();
 		return res.status(400).json({ error: "Invalid request!!" });
 	}
 }
